Remove projectile-enable-caching t
This was unnecessary, and made it so that 1. navigating to a new file
was slower since it would add it to cache and 2. after it was added to
cache it would have the wrong \ vs /, which made it so the file would
appear twice when selecting for a file, since projectile looks at the
per-file cache as well.